refactored all parsing code into one place
